Feat: bug relative to the first grid card fix
I had problem to manage the way to display the grid cards with
the "Your project" title without to make it a part of the grid.
So I applied a position relative to it in order to remove it to
the document flow and manage it without being part of the grid.